collab: Factor out subscription kind determination
